add ArgoCD
* confirm ArgoCD login page
* need to configure ArgoCD admin
* update sh files

+++ redeploy_django.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 
-kubectl delete deployment django -n final
+if $(kubectl get deployment django -n final); then
+    kubectl delete deployment django -n final
+fi
 
-kubectl delete secret django-secret -n final
+if $(kubectl get secret django-secret -n final); then
+    kubectl delete secret django-secret -n final
+fi
 
-kubectl delete configmap django-config -n final
+if $(kubectl get configmap django-config -n final); then
+    kubectl delete configmap django-config -n final
+fi
 
-kubectl apply -f manifests/
+kubectl apply -f ./manifests
